Blur detection for browser
