+++ demo.html
@@ -15,6 +15,51 @@
         </form>
         <p id="blur_score"></p>
         <canvas id="canvas"></canvas>
-        <script src="measure_blur_browser.js"></script>
+
+        <script src="dist/measure_blur.js"></script>
+        <script>
+          function readImageFile(rawFile) {
+            return new Promise(function(resolve, reject) {
+              if (!rawFile) return;
+
+              var reader = new FileReader();
+              reader.onload = function(readerEvent) {
+                var img = new Image();
+                img.onload = function() {
+                  resolve({ // NOTE: This is not ImageData object!
+                    rawFile: rawFile,
+                    data: img,
+                    width: img.width,
+                    height: img.height
+                  });
+                };
+                img.onerror = reject;
+                img.src = readerEvent.target.result;
+              };
+              reader.readAsDataURL(rawFile);
+            });
+          }
+          function handleImageInput(e) {
+            var done = function(img) {
+              var canvas = document.querySelector('#canvas');
+              var context = canvas.getContext("2d");
+
+              canvas.width = img.width;
+              canvas.height = img.height;
+              context.drawImage(img.data, 0, 0);
+
+
+              var score = measureBlur(context.getImageData(0, 0, canvas.width, canvas.height));
+              console.info('Blur score: ' + Number((score.avg_edge_width_perc).toFixed(2)));
+              console.log('Detail blur score:', score);
+            };
+            var fail = console.error;
+            readImageFile(e.target.files[0]).then(done, fail);
+          }
+
+          document
+            .querySelector('#upload_image')
+            .addEventListener('change', handleImageInput, false);
+        </script>
     </body>
 </html>
