+++ measure_blur.js
@@ -1,7 +1,4 @@
-var fs = require('fs'),
-    Canvas = require('canvas'),
-    Filters = require('canvasfilters').Filters,
-    image;
+var Filters = require('canvasfilters').Filters;
 
 function detectEdges(imageData) {
     var greyscaled, sobelKernel;
@@ -90,37 +87,8 @@ function detectBlur(pixels) {
     };
 }
 
-    
-function createImage(error, data) {
-    if (error) {
-        console.error('Unable to read image file!');
-        throw error;
-    }
-    image = new Canvas.Image;
-    image.onload = drawImageOnCanvas;
-    image.src = data;
-}
-    
-function drawImageOnCanvas() {
-    var canvas = new Canvas(),
-        context;
-
-    canvas.width = image.width;
-    canvas.height = image.height;
-    context = canvas.getContext('2d');
-    context.drawImage(image, 0, 0);
-
-    showBlurScore(context.getImageData(0, 0, canvas.width, canvas.height));
+function measureBlur(imageData) {
+    return detectBlur(reducedPixels(detectEdges(imageData)));
 }
 
-function showBlurScore(imageData) {
-    stats = detectBlur(reducedPixels(detectEdges(imageData)));
-    console.log('Blur score:', Number((stats.avg_edge_width_perc).toFixed(2)));
-    console.log(stats);
-}
-
-if (process.argv.length >= 3) {
-    fs.readFile(process.argv[2], createImage);
-} else {
-    console.info('Usage: node measure_blur.js path/to/image/file.jpg');
-}
+module.exports = measureBlur;
